Plane: create afs_plane.h to match copter and rover structure
